Updated code to represent new wishes.
Changes modified table to extended table;
Updated API/jira parent key code;
Updated queries

+++ Models/DataEmailEmail.cs
@@ -30,7 +30,7 @@ public partial class DataEmailEmail
 
     public int? ThreadId { get; set; }
 
-    public long? WordCount { get; set; }
+    public int? WordCount { get; set; }
 
     public virtual ICollection<ResultArchEmailsAllIssue> ResultArchEmailsAllIssues { get; set; } = new List<ResultArchEmailsAllIssue>();
